pyproject.toml
[tool.poetry]
name = "llmstudio"
version = "0.3.2"
description = "Prompt Perfection at Your Fingertips"
authors = ["Cláudio Lemos <claudio@tensorops.ai>"]
license = "MIT"
homepage = "https://llmstudio.ai/"
repository = "https://github.com/tensoropsai/llmstudio"
documentation = "https://docs.llmstudio.ai"
readme = "README.md"
keywords = ["ml", "ai", "llm", "llmops", "openai", "langchain", "chatgpt", "llmstudio", "tensorops"]

[tool.poetry.urls]
"Bug Tracker" = "https://github.com/tensoropsai/llmstudio/issues"

[tool.poetry.dependencies]
python = "^3.9"
pydantic = "^2.5.3"
requests = "^2.31.0"
fastapi = "^0.108.0"
uvicorn = "^0.25.0"
openai = "^1.6.1"
tiktoken = "^0.5.2"
python-dotenv = "^1.0.0"
tokenizer = "^3.4.3"
aiohttp = "^3.9.1"
sqlalchemy = "^2.0.27"
anthropic = "^0.16.0"

[tool.poetry.dev-dependencies]
pytest = "^7.4.4"
pytest-mock = "^3.12.0"
pytest-asyncio = "^0.23.3"

[tool.poetry.scripts]
llmstudio = "llmstudio.cli:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
